AR-19 done. Also updated init.sh

+++ db/init.sh
@@ -49,12 +49,17 @@ if [ $# -eq 0 ] || [ "$1" == "new" ] || [ "$1" == "all" ]
         PGPASSWORD=${password} psql -U ${username} -p ${port} -h ${hostname} ${database} < ./Starred_Resumes/Starred_Resumes.sql;
         PGPASSWORD=${password} psql -U ${username} -p ${port} -h ${hostname} ${database} < ./Pending_Requests_Sent/Pending_Requests_Sent.sql;
         PGPASSWORD=${password} psql -U ${username} -p ${port} -h ${hostname} ${database} < ./Job_list/Job_list.sql; 
+        PGPASSWORD=${password} psql -U ${username} -p ${port} -h ${hostname} ${database} < ./Experiences/Experiences.sql; 
         } > /dev/null 2> "$logfile"; [ -s "$logfile" ] || rm -f "$logfile"
         if [ -f "$logfile" ]
             then
-            echo "An error exists, please check $logfile for detailed errors"
-            else
-            echo "Database created successfully"
+              if grep ERROR "$logfile"
+                then
+                  echo "An error may have occurred while creating tables, please check $logfile for detailed errors"
+                  echo "If you believe this is an error on our side, delete the log file then proceed once again"
+              else
+                echo "Database created successfully"
+              fi
         fi
     else
         echo "$config not found."
@@ -64,27 +69,31 @@ fi
 
 if [ "$1" == "dummy" ] || [ "$1" == "new" ] || [ "$1" == "all" ]
     then
-    #     if [ -f "$logfile" ]
-        # then
-            # echo "An error, please check $logfile for detailed errors before proceeding"
-            # echo "If you believe this is an error on our side, delete the log file then proceed once again"
-        # else
-            echo "Inserting dummy data"
-            . $config
-            { PGPASSWORD=${password} psql -U ${username} -p ${port} -h ${hostname} ${database} < ./dummy_data/jobseeker_data.sql
-            # Make sure company is created first and employer info has existing company_id
-            PGPASSWORD=${password} psql -U ${username} -p ${port} -h ${hostname} ${database} < ./dummy_data/company_data.sql
-            PGPASSWORD=${password} psql -U ${username} -p ${port} -h ${hostname} ${database} < ./dummy_data/employer_data.sql
-            PGPASSWORD=${password} psql -U ${username} -p ${port} -h ${hostname} ${database} < ./dummy_data/job_data.sql
-            PGPASSWORD=${password} psql -U ${username} -p ${port} -h ${hostname} ${database} < ./dummy_data/resume_data.sql
-            } > /dev/null 2> "$logfile"; [ -s "$logfile" ] || rm -f "$logfile"
-            if [ -f "$logfile" ]
+         if [ -f "$logfile" ]
+         then
+            if grep ERROR "$logfile"
                 then
-                echo "An error exists, please check $logfile for detailed errors"
-                else
-                echo "Dummy data inserted correctly"
+                  echo "Please check $logfile before inserting dummy data"
+                  echo "If you believe this is an error on our side, delete the log file then proceed once again"
+            else
+                echo "Inserting dummy data"
+                . $config
+                # ORDER MATTERS due to references in the tables.
+                { PGPASSWORD=${password} psql -U ${username} -p ${port} -h ${hostname} ${database} < ./dummy_data/jobseeker_data.sql
+                PGPASSWORD=${password} psql -U ${username} -p ${port} -h ${hostname} ${database} < ./dummy_data/company_data.sql
+                PGPASSWORD=${password} psql -U ${username} -p ${port} -h ${hostname} ${database} < ./dummy_data/employer_data.sql
+                PGPASSWORD=${password} psql -U ${username} -p ${port} -h ${hostname} ${database} < ./dummy_data/job_data.sql
+                PGPASSWORD=${password} psql -U ${username} -p ${port} -h ${hostname} ${database} < ./dummy_data/resume_data.sql
+                PGPASSWORD=${password} psql -U ${username} -p ${port} -h ${hostname} ${database} < ./dummy_data/experiences_data.sql
+                } > /dev/null 2> "$logfile"; [ -s "$logfile" ] || rm -f "$logfile"
+                if [ -f "$logfile" ]
+                  then
+                    echo "An error may have occurred while inserting dumy data, please check $logfile for detailed errors"
+                  else
+                    echo "Dummy data inserted correctly"
+                fi
             fi
-        # fi
+         fi
 fi
 
 if [ "$1" == "clear" ]
